requirements.txt
Django>=5.2.4
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
django-filter>=23.5
psycopg2-binary>=2.9.9
whitenoise>=6.9.0
gunicorn>=21.2.0
celery>=5.5.3
redis>=6.2.0
django-redis>=6.0.0
django-celery-beat>=2.8.1
django-celery-results>=2.6.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
backoff>=2.2.1
dj-database-url>=2.0.0
